// src/lib/api/models/MonthlyRevenue.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MonthlyRevenue = {
    year?: number;
    month?: number;
    monthName?: string;
    revenue?: number;
    eventCount?: number;
    registrationCount?: number;
};